Ensure Paranoia runs after_commit on: :destroy callbacks
Rails 5.1+ adds a @_trigger_destroy_callback instance variable to
ActiveRecord's internal code, which it checks to see if it should run
those callbacks. Paranoia 2.4.1 isn't aware of that variable, which
causes after_commit on: :destroy callbacks to not run for soft-deleted
records.

Since Paranoia is no longer getting updated, we're forced to fork it
and add awareness of the new variable ourselves. The changes in this
commit are based on those in https://github.com/rubysherpas/paranoia/pull/404/files .

The main difference from that source is that since we know this private
fork is only getting used by us, I can drop the "are we on Rails 5.0
or earlier?" conditionals, though they should be added back if we ever
intend to submit this upstream.

Since I dropped those conditionals I changed the gemspec to require
ActiveRecord 5.1 or greater.